Moved inputsystem to input package
